move variables out of main to prepare for refactoring
Some variables are shared between the code, but it's to many variables
to have them as return values from Init().
Alternatively, some object can be made to hold them or modify pointers.

The problem for refactoring is runtime.LockOSThread().
Because of the way go handles threads and function calls,
the variable initializers cannot simply be moved to an init function.